Update "find a pizza I can eat" function
The function was able to pass the test due to setting the test length of productsICanEat to two but that number should be one. Originally returned two pizzas by checking for a pizza that contains nuts && does not have mushrooms. Now it returns one pizza that does not contain nuts && does not have mushrooms. Also changed the length that we test for.